physionet.urls: enable automated URL testing.
We want to skip lightwave tests if sandboxed-lightwave isn't
installed, so we need to add an exception in TEST_CASES to handle that
case.  In order for that to work, we need to give a name to the
formerly nameless URL pattern "cgi-bin/lightwave".

The parameter values defined in TEST_DEFAULTS refer to test data
defined in demo-project.json, demo-user.json, and demo-physionet.json.