search.urls: enable automated URL testing.
There are no demo "legacy challenge" projects defined (which perhaps
there ought to be), so we can't test the redirect_challenge_project
view.  In order to skip it, we need to give a name to the formerly
nameless URL pattern.

The parameter values defined in TEST_DEFAULTS and TEST_CASES refer to
test data defined in demo-project.json and demo-user.json.