users_aws_access_list_json: new view.
/console/users/aws-access-list.json provides a table of projects, and
AWS IDs for users who currently have access to each project.  This
view requires 'user.view_user' permission, which might not be exactly
what we want.

This is a temporary kludge to support an upcoming event, and will
likely be removed or substantially redesigned in the future; don't
rely on it.

Co-authored-by: Chrystinne Fernandes <chrystinne@gmail.com>