sso.tests.test_views: a User must have a Profile.
Every User must have a Profile, which is created by the
RegistrationForm - if there's a way in the system to create a User
without creating a Profile, that's a bug.  So if we manually create a
User we need to manually create a Profile as well.

It's probably not a great idea to call User.objects.create by hand,
and it would be better for tests to instead use the demo data
(demo-user.json, etc.) if possible.