process_invitation_response: perform operations as a transaction.
There are several changes to make here (updating AuthorInvitations,
creating an Author, creating an Affiliation) and if something goes
wrong we don't want to have done some of those things and not others.

Also, rearrange the code to reduce nesting and make it clear that we
are only going to process one InvitationResponseForm.