process_invitation_response: take affiliation data from form.
When accepting an invitation to coauthor a project, instead of copying
the affiliation from the Profile (author.import_profile_info()), use
the affiliation that is entered in (and required by) the
InvitationResponseForm.

Instead of returning a tuple to indicate whether the profile contained
an affiliation, simply return the project or None.  Note that this
function was previously broken, and would return None (where the
caller expected a tuple) when the form was not valid.