project_home.html: remove unnecessary wording about affiliations.
Previously, accepting an authorship invitation would automatically
copy the affiliation from their profile, so some additional text was
included to remind people that they could change it, or remind them
that they needed to fill it in.

This is now unnecessary since the form includes an affiliation field
to make this explicit.