Refactoring of GenerateSignedUrl(file upload process)
Context: Currently the file upload process allows any user who has access to admin console to upload the files. This is a security issue. We need to restrict the file upload process to only the users who have access to  either of the following roles: 1. Admin 2. Super Admin 3. Super User
1. Project Author
2. Project Editor
3. System Admin

My change will restrict the access to unauthorized users. i also added a test case that checks and verifies the access to the file upload process.