Updated the File Upload View to make it consistent with the file upload process in the UI + minor bug fix
Context1: This is a follow up to pr #1721. Currently, the file upload view doesnot check the project submission status and allows the user to upload the file even if the project is not in the submission phase where upload is permitted.
The change aims to increase the security of view in case a user with access to the platform tries to maliciously upload a file to the project.

Explanation:
On the UI, Authors can only upload files if the project is not submitted or in revision state.
and Editors can only upload files if the project is accepted.

But this is not the case with the file upload view. Before this change a malicious user can bypass the UI and upload files by sending the request directly.(See TRA report for more details)

Context 2(minor bug): It seems sometime, django filters with return duplicate objects when we have or filter, Q and one to many relation
This causes the file upload view to throw a `MultipleObjectsReturned` exception on the get_object_or_404 call.So adding a distinct() call to the queryset fixes the issue.
Here is a related issue on stackoverflow: https://stackoverflow.com/questions/34468505/why-is-django-queryset-with-q-expression-returning-duplicate-values

To replicate the bug, please call the generate_signed_url view for project ` Demo software project` by user `tpollard@mit.edu`

Note: Although it would have been a good idea to open 2 PR's, i did both changes in single PR because on of the test case(test_invalid_access) i added will fail with out the bug fix