Fixed Error caused when using distinct and order_by with sqlite
Explanation:

Django seems to throw error when we use distinct on queries and  order_by is used(unless we use the same column in distinct and order_by). This is a bug in Django. This patch fixes the bug.

In our case it looks like get_object_or_404(where we are providing a queryset with distinct()) calls a function that does somekind of ordering. So we were getting the error.

Here is the link to our test failure https://github.com/MIT-LCP/physionet-build/actions/runs/3593868625/jobs/6051387850

So i think adding order_by and distinct on the slug column should prevent the error from happening