Updated the workflow to avoid the errors that came with use of distinct()
Context:
I tried to fix the error(link is attached below) that appeared when using distinct() in queryset. I couldn't fix it/narrow its cause, so i changed the workflow to avoid the error entirely.
At the end the view works same as before(with 1 improvement, now in case the user is not authorized we will throw 403 error, earlier for all the cases such as project not found, unauthorized user we used to send 404 error)
https://github.com/MIT-LCP/physionet-build/actions/runs/3593868625/jobs/6051387850

Explanation:
Earlier, we were listing all the projects the current user has access to as either author or editor, and then trying to find the project.
Now we are first checking if project exists, if it exists, then we check if the current user has access to it.

To make sure, this change doesn't break anything, i tested by running the test, and also tried to come up with scenarios that this change might cause a bug(i couldn't think of any)