Updated code to use permission instead of group name, and use model methods to fetch Project related details
I opted out of adding a function that checks if the current user can edit the project because i think that function is not very useful to file upload process as uploads are different compared to editing abstract, or other project details.
For example for uploads only submitting author can upload files, but outside of uploads, any author can edit the project.

Also i opted out of making changes to project_auth, because i have another PR which is changing the project_auth to work with the new permission we added. So it would probably get confusing to merge or maybe the tests will fail when we merge the two PRs.