Rebasing the PR to dev
Added a new permission to determine who gets to be assigned as a project editor and Supporting Changes

Context:
Previously anyone who had is_admin flag was able to be assigned as project editor.  Following #1719, we are removing the use of is_admin flag.
This permission are assigned to relevant groups, and will be used to determine who can be assigned as project editor.
If for a  deployment of PhysioNet has a different set of groups(or groups have been), they just need to add the new permission to the relevant groups.

Changes
1. Added a custom permission on activeproject `can_edit_activeprojects`
2. Fixed the forms to check for the users who have this permission. Only those users can be assigned as project editors.
3. Fixed test cases as with the test, the old fixtures were not working(as the logic of who can be assigned as project editor has changed). So, I have updated the fixtures to add a new user with correct permission, and updated an existing user with permission.
4. Updated fixtures to Provide both Editor groups ability to be assigned as project editors.