Fixed failed test case
After the code was merged with another PR, the test case was failing because the user tom was added to Admin group, So the test was failing as the test was supposed to check if editors can edit the project. So I have changed the user tom to editor and the test case is passing now.
I fixed it by changing the project to another one with another user where the user is not in the admin group. Hence the test was able to test its functionality without any issues.