NewProjectVersionForm: set correct corresponding_email.
Each Author has corresponding_email, which should be a verified email
address belonging to that person.  If this person is selected as the
corresponding author of the project, their corresponding_email will be
the one stored in the project Contact; if they are not selected as the
corresponding author, their corresponding_email is ignored.

In particular, it is very wrong to set every author's
corresponding_email to the primary email address of the person who is
submitting this form (= the submitting author).  The submitting author
may not be the corresponding author, and if they *are* the
corresponding author then their primary address may not be the one
they want to advertise publicly.

When creating a new project version, the new corresponding author will
be (by default) the same as the original corresponding author.  If
their old corresponding_email is still valid, that should be carried
forward.  If they have the same address modulo capitalization, that
counts too.  Otherwise, select their public email address if any, and
select their primary email address as a last resort.