Creating a new permission that will decide if the user can be an project editor
The current setup allows anyone with is_admin permission to be assigned as editor. Since we have refactored the code to remove the use
of is_admin, we need to create a new permission that will decide if the user can be an project editor.

This permission is to be assigned to the user by the Admin either through permission group or directly to the user.