Created a method that can return a list of users with a given permission
Looks like django doesnot have an easy way to get a list of users with a given permission. So I created a method that can return a list of users with a given permission.
This will be used to fix the form that we use to assign editors to project on console app

Since this method is something that is independent of the User model, i created it as staticmethod on the User model. This way we can call it without having to create an instance of the User model.