physionet-build-test.yml: display context for flake8 errors.
For each error message reported by flake8, display the actual source
line it is complaining about, along with the two preceding lines.

In particular, it's sometimes confusing to see line numbers coming
from the auto-merged tree.  Showing the actual source should make it
easier to see what the problem is.