Making the number of associated emails per account configurable per user if needed.
The Goal of this fix is to limit the number of associated emails per account. This is done by adding a new field to the user model called associated_emails_allowed.
This field is set to 10 by default(which comes from the .env file). If the user needs more associated emails, they can contact the admin and ask for more. The admin can then change the value of this field to the desired number.

Although, i can understand that modifying the user model or any models is probably not desirable in most cases as it would make changes to database. I will leave it to the reviewers to decide if this is the best way to go about this.

If we don't want to make it configurable per user, i can remove the associated_emails_allowed field and just make it a constant in the settings file.