Fixing the failing test case caused due to all django fields having not null criteria
The upgrade_test.sh failed due to the not null criteria. So my guess is that the previous state of database didnot have our new field and the migration was applied, but weirdly the default value was not set.
So i added null=True in the IntegerField to see if my suspicion is correct.