Removing the changes in model
After discussion with team, we dont want to make changes to the user model/database. So removing the changes in model and adding a new migration file to remove the field from database.
if user needs the limit increased, then the admin can simply increase it in the config file, and all users will have their limit increased.