Added a utility function to verify if the uploaded file is a pdf file or not.
Currently users are able to upload any file type by changing the file extension in the training form.
This commit adds a utility function to verify if the uploaded file is a pdf file or not.
Looks like we are already using pdfminer.six to extract text from pdf files.

So I have used the same library to verify if the uploaded file is a pdf file or not.

This function is used by the view that handles the training certificate upload. The pdf validation is done using the utility function