Update test to work with the training pdf validation
The new change makes sure that the file uploaded is a valid pdf and is readable.
Our previous test was using django core SimpleUploadedFile to create a plain/txt file and uploading that which is not a valid pdf(even if we change the content type).
So the solution to this would be to either create pdf on the fly during the test or use a static valid pdf file.

I chose the later option, because creating pdf on the fly would require us to add a new dependency to the project which might introduce a security risk and need more discussion.