Fix for flickering for edit_email
Please check the issue on what caused the flickering and how we might be able to fix it.
https://github.com/MIT-LCP/physionet-build/issues/1764

In Summary, the flickering was caused by the modal being present inside a li tag with class `list-group-item`. The value of z-index on hover, was causing the flickering. The fix was to remove the z-index on hover.

The simple fix was to override the z-index on hover for the modal. We can do this by adding the following to the physionet.css file.
This should make sure, that the override is applied throughout the site(as the physionet.css is loaded in base_css.html which is loaded in base.html)