Refactor the wrapper function to use prevent any possible override
On PR: https://github.com/MIT-LCP/physionet-build/pull/1763 , i added a wrapper function to remove use of the private function
On the PR i had named the function `generate_signed_url` which ironically was also name of the view function on the project/views.py

So, i have renamed the wrapper function to `generate_signed_url_helper` to prevent any possible function override.

Strange that the test didn't catch this.