getmigrationtargets: fail if there are conflicting leaf migrations.
If there are multiple leaf nodes for the same app in the migration
graph, the migrate command will fail (as will makemigrations --check).
This is an error that has to be corrected (typically by rewriting the
migration graph) before the new migrations can be installed.
Therefore, getmigrationtargets should fail in this case.