settings: new variable GOOGLE_APPLICATION_CREDENTIALS.
This variable should be set to the absolute path of the Google
application credentials file (determined by the
GOOGLE_APPLICATION_CREDENTIALS variable in .env); or None if there is
none.  This is needed in order to disable optional Google integration
features when there is no Google service account defined.

(This variable is also exported into os.environ for some reason,
probably because other packages are relying on it.)