manage_published_project: handle absent GOOGLE_APPLICATION_CREDENTIALS.
If the GOOGLE_APPLICATION_CREDENTIALS variable isn't set, this should
be treated as "no Google service account is available."  This should
be determined by the settings module and not by os.environ.  (We also
don't want the system to crash if the variable is undefined, which it
is by default.)

Moreover, we don't want to check the filesystem on every request; just
say that if the variable is defined then the Google integration
options should be displayed on the page.  If the variable is defined
and the file doesn't exist, this is a configuration error (and should
result in an error if you try to click the button) but shouldn't be
silently hidden.