getmigrationtargets: print "zero" to signify no migrations.
If an app has just been installed, that app has no migrations applied
and no models in the database.  Ordinarily, the first migration in the
app (0001_initial) will create some models, which means there's no
named migration corresponding to the no-models-defined state.

Instead, the special name "zero" can be passed to the 'migrate' or
'migratetargets' command, to instruct it to unapply all migrations for
the specified app.

test-upgrade.sh tries to test that all migrations are reversible, by
checking 'getmigrationtargets --current' before applying migrations
and again after unapplying them.  (The reason for this check is that I
don't entirely trust the one-app-at-a-time, named-target-migration
approach to correctly handle complex cases with branching dependencies
and inter-app dependencies.)

Although in practice it's unlikely that we'd ever want to reverse the
0001_initial migration, it's good to be consistent.  In order for this
to work, 'getmigrationtargets --current' must explicitly list "zero"
as a target for apps with no migrations currently applied.

(In the very bizarre situation where the initial migration is
MIGRATE_AFTER_INSTALL, 'getmigrationtargets --early' should likewise
print "zero".)

The ordering of "zero" targets shouldn't matter (because by definition
they can't depend on each other), so in case two apps are added at
once, they're listed in alphabetical order for reproducibility.

We don't list "zero" targets for apps that don't have any migrations
at all; this allows apps with no models to be freely added and removed
without impacting the database, and (possibly, though this is
untested) provides a path for uninstalling an app that previously
contained models.