feature: add granted_access bool used for controlling notification to user
The grant_aws_open_data_access and grant_gcp_group_access utility functions
return a message if they succeed or fail. The project view was checking this
message against custom strings to determine whether to provide a success or
error modal to inform the user. This commit changes these functions and the
view to use a boolean variable which seems less fragile.