feature: e-mail user rather than raising server error if gcp data access fails
Currently, if the system is unable to grant access to a user e-mail to GCP,
i.e. unable to add their cloud e-mail to a Google Group, the server will
raise an error. Instead, it would be better to notify the user, and provide
some advice on what the cause of the error may be.