feature: provide more useful error message for HTTP 412 response from google
Google Admin sometimes returns a 412. It's not 100% clear, but this seems
to happen if the server tries to add a non-Google e-mail to the group.
Typically, this occurs if the user has added a non-Google e-mail in
their cloud profile. This commit adds a notification to hint the user
towards a solution.