Event.enroll_user: search for permission by app label.
When a user is enrolled in an event, a user permission is added which
causes the "Events" menu to appear on their navigation bar.

This permission is identified by the codename "view_event_menu";
however, there's no guarantee that this codename isn't also used by
some other Django app.

(There's also, as far as I know, nothing preventing two models in the
same app from using the same codename, but at least all the models for
one app are defined in the same Python module, and lots of Django code
assumes codenames are unique within an app.)

To ensure we are referring to the correct permission object, filter by
the app label as well as the codename.