get_users_with_permission: add required app_label argument.
In general, there's no guarantee that the same permission codename
isn't used by two or more Django apps.  To ensure that we are
referring to the correct permission object, filter permissions by the
app label as well as the codename.

(There's also, as far as I know, nothing preventing two models in the
same app from using the same codename, but at least all the models for
one app are defined in the same Python module, and lots of Django code
assumes codenames are unique within an app.)