Fix IntegrityError on duplicate event title
Currently, when we edit/add an event to have a title that already exists, we get  `IntegrityError at /events/. UNIQUE constraint failed: events_event.title, events_event.host_id)`.
This is because we are not checking for the uniqueness of the title in the form. This PR fixes this issue by overiding the clean method of the form.

One thing to note is that, the behaviour of the website is slightly different when there is an duplicate title condition

1. For add event
    When someone submits a form with a duplicate title, the form is re-rendered with the error message. The user can then edit the form and submit again.

2. For edit event
    When someone submits a form with a duplicate title, the users is redirected to the event home page error message. The modal is not re-rendered with the error message.
    This is because, to have the same behaviour on edit event, we would need to add an input for slug to edit event form(with javascript) and then read that slug value in view through request.POST.get('slug') and use that to find the event in database

    The reason i didnot do that is because when we access the data before form validation, it might lead to sql injection attack vulnerability. So, i decided to redirect the user to the event home page with the error message.