Fix bug in edit event form which prevented the editing other details except title of same event
Before the changes, the logic that checks for duplicate title + host was same for both create and edit event form. This was causing the edit event form to fail if the user tries to edit any other details of the event except the title.
So i added a additional checks that also compares the title of the event with the title of the event that is being edited. This way the logic will not fail if the user tries to edit any other details of the event except the title.
i.e if the current title is same as the instance title, then the code will throw an error.