Fix calling GoogleCloudStorage property method
On the previous PR, i was calling the `bucket` method  of 'GoogleCloudStorage` as a method `storage.bucket()` instead of  `storage.bucket`.
On the `GoogleCloudStorage` class, the `bucket` method is a property, so it should be called as `storage.bucket` instead of `storage.bucket()`

Thanks to Karol for pointing this out. I didnot catch this earlier because i couldnot test the gcp upload code locally and had no idea about how the `property` decorator works.