Tests: only the requester should view data access requests.
The data_access_request_status_detail page is provided for data access
*requesters* to view their past applications and status.

This page should only be accessible to the requester themselves (other
URLs are provided for the use of the reviewer), although currently,
the function does not limit access beyond requiring a user login.

Correct this test case so that we can enforce the appropriate access
restriction in the view.