.github/workflows: mark git repository as a safe directory.
Modern versions of git will complain if trying to operate on a
repository directory not owned by you (which is, indeed, unsafe.)

In the GitHub workflows, this error is triggered for some reason,
maybe because of something wrong in how the container is set up or
maybe because of something weird about how "actions/checkout" works.

Regardless of the reason, there doesn't seem to be any reason not to
mark the repository directory as safe, and doing this should be
harmless if GitHub later fixes the issue.