Making AutoRejection optional and turn it off by default
Added a new boolean on configuration file to make the auto-rejection of pending credentialing applications optional.
The idea is that in case, if someone wants to disable the auto-rejection of pending credentialing applications, they can do so by setting the value of the new boolean to False.
Similarly, for existing deployment like healthdatanexus.ai, they might not want the auto rejection feature, so it seemed like a good idea to make it optional and turn it off by default.