Adding a test case to check management command
I made a copy of `test_accept_pending_requests.py` and modified it to test the `reject_pending_credentialing_applications` management command.

Quick note: The django recommendation to test the management command is to use `call_command` and not to test the `handle` method directly. https://docs.djangoproject.com/en/4.1/topics/testing/tools/#management-commands

I think this is because the `handle` method is not intended to be called directly and since the  goal of the command should be to test the command itself, not the `handle` method.

However, on this commit, i am testing the `handle` method directly, just in case there is some reason that we did it this way in the past.