test_urls: treat redirection to /login/ as an error.
Many views use django.contrib.auth.decorators.login_required or
similar, which handles authorization failures by redirecting the
client to the login page (settings.LOGIN_URL).

When we are testing views with TestURLs, authorization failures should
be considered an error; the purpose of TestURLs is to see what the
page looks like when viewed by an appropriately-authorized client.

On the other hand, we do want to allow views to return redirections in
general (for example, the published_project_latest view always returns
a redirection.)

So, treat redirection to settings.LOGIN_URL as a special case, which
(like a 403 response) should be considered a test failure.