Added a new if else block to make it easier to follow code changes
So i decided to add a new if else block. This if else block is the end goal for this Pull Request, and i will work towards refactoring the existing template,
removing the code from the (if else block for credentialed and open access stuff) and eventually removing the old if else block.

After lots of commits following this one, at the end we should end up with all the code from the (if else block for credentialed and open access stuff)
moved to the new if else block and the old if else block removed.