Refactor if block for CONTRIBUTOR_REVIEW
This if block was originally under the if block project.access_policy,

And the behaviour now should be same as before, i.e if the project.access_policy is AccessPolicy.CONTRIBUTOR_REVIEW, then show the content under the if block.
After this change, the `has_access` will be true if the project.access_policy is AccessPolicy.CONTRIBUTOR_REVIEW, so this if block should also get executed. in case project has access_policy
project.access_policy == AccessPolicy.CONTRIBUTOR_REVIEW