user/urls.py: add parameters for TestURLs.
Most URLs in this app deal with the user's own profile and settings,
which requires a logged-in user; use aewj as the test subject.

A few views require dynamically-generated (time-sensitive) tokens;
there's currently no way to test these using TestURLs, so they are
skipped.

There are currently no test cases for the auth_orcid view;
implementing such a test case requires mocking the remote oauth
server, which is also beyond the capabilities of TestURLs.