Fixed template to show the Event Menu to regular users once they have the view_event_menu permission.
Looks like when we refactored the events app and made it separate from the user app. we forgot to change the `perms.user.view_event_menu` to `perms.events.view_event_menu` in the template.
I changed that and now users can see the events menu after they have requested to join an event.