Record autorejection in database
1. As pointed out by @bemoody, we need to record the reason for autorejection in the database. This is a simple straightforward change to the model.
As per @bemoody's suggestion, I've added a class that can hold the reasons for autorejection, and a field to the CredentialApplication model to hold the reason.
Currently, the only reason is that the application has been pending for too long, but we may want to add more reasons in the future. The current implementation allows for that in the future with little changes.

2. In addition to that we also wrapped the auto rejection in a separate method