Limiting the total number of credentialing applications that can be rejected in a single run to a default of 5
The total number of applications that can be rejected in a single run can be overridden by passing the --number option to the command.

As suggested by @bemoody on https://github.com/MIT-LCP/physionet-build/pull/1846, we shouldnot send a large number of rejection emails in a single run(this might cause issues with the email server).

The idea is to run the management commands multiple times a day(perhaps once every hour). This should be fine as rejecting the applications after the deadline is not a critical task and rejecting applications by few hours should not make a difference.