Set the auto reject management command to run every 1 hour
Since we are now rejecting only 5 applications at a time, we will need to run the command more frequently.(as we might have huge backlog of applications and if we just do 5 applications per day, it will probably takes months)
So we will run the command every 1 hour. we should be able to clear backlog 24 * 5 = 120 applications per day.