Fix failing test case
The test was failing because by default the auto rejection is disabled. But the test case was expecting the auto rejection to be enabled.(as the test was written before the auto rejection feature was made optional)

So i updated the test to check if the auto rejection is enabled or not. If it is enabled then the test case will pass otherwise it will be skipped.(if we want to run the test case on CICD we need to turn on the autorejection by default)

In addition to that i also removed redundant code from the test case and the management command by creating a new method and using that.