Slight refactoring of the event registration page
As discussed with @alistairewj, we want to use the same page to show event details, register participants to event, and everything related to the event details.

To start that it makes sense to rename the views, templates as the current view says `event_add_participant`, which is not a good name for a view that shows event details(and does other things).

So i started refactoring the viewname, template name and small things in the template

Other commits will follow to refactor the view and template to show event details and add other features.