Cleanup of the event registration page(now event detail page)
Context:
Previously on the event registration page, we didnot allow the host to visit the registration page as it didnot make sense to register for an event that you are hosting.
But since we now converted the event registration page to event detail page, we need to allow the host to visit the page so that they can see the details of the event.

Change: The change removes the limitation on view, which didnot allow the host to visit the event detail page(Previously event registration).