Fixed Bug in test_reject_pending_credentialing_applications.py
Problem:
When the test was first written, the assumption was that when the management command was run, it would reject all pending credentialing applications and validate that there is no more applications remaining to be rejected.
However, later on, the logic was changed to reject only x number of pending credentialing applications at a time.
So, if we have 10 pending credentialing applications, and we run the management command to reject 5 pending credentialing applications, the test would fail because there are still 5 pending credentialing applications remaining.

Solution:
We really don't need to validate the remaining number of applications, we just need to validate that the selected number of applications were rejected. So, I removed the validation of the remaining number of applications.

Why i didn't catch this earlier?
It was because on the user-demo.json fixture, there was only 1 pending credentialing application. So, when I ran the test, it would reject the 1 pending credentialing application and validate that there are no more pending credentialing applications remaining. So, the test would pass.

I did check the code once again and couldnot find any other issues with the test or the management command.