Refactoring event detail page to manage event participation access easily
On this commit, i created a block of code on event/views.py.
In this block, we are checking if the current user should be allowed to register to the event.(To decide if the user should see the event participation form or not i refactored the previous logic )

If the user is not allowed to register to the event for whatever reason, we will hide the form, and show appropriate messages.

This should allow us to easily prevent/close registration in case
1. Event has Ended
2. User is already on the waiting list/registered
3. User doesnot have the allowed domain email address
4. Similarly in future if we decide to add more conditions to prevent registration like a limit on how many people can register for an event, or a registration deadline, etc.