Update EventParticipant model to support co-hosts
For Events app, we want event hosts to be able to make some of the EventParticipant as co-hosts so that they can approve participation requests and manage the event.

For this, i thought it might be a better idea to add a new field to the EventParticipant model to indicate if the participant is a co-host or not.
(I suppose we could have a separate model for co-hosts but i think it might not be necessary as we don't need to store any additional information for co-hosts)