published_project: add project version number for wfdb-python.
Current versions of wfdb-python attempt to parse the project version
by looking for a line in the HTML file which contains the string
"Version:" followed by the version number.  (See the get_version
function in wfdb.io.download.)

This parsing is very fragile but in order to avoid breaking a lot of
existing wfdb-python scripts, we should try to keep this working for
the time being.

To reduce the possibility of false matches, the version string is
placed close to the top of the HTML file (fortunately, wfdb-python
should ignore matches after the first.)