Provide the Event Host access to Admin console
Because historically we have been using Admin Console for stuff like creating DUA, license and the Event Agreement is kind of same, it makes sense to put the code behind admin console. So for that the first step is to provide the `Event Host` access to Admin console.
This should not pose any security risk as all the other views or links in templates are secured by different permission check.

So after this change, the `Event Host` should be able to admin console and only see the `Event Participant stuff`(The code for that is coming after this commit)