bug fix: to show the registration status to participants only
Bug description:
There was a bug where host could see the message "On Waiting List", which doesnot make sense as hosts dont need to register to the event.

Solution:
I added a new template tag that checked if the current user is on the waiting list for the particular event.
And used that tag on the event_home.html template. Now we are making sure that the user is in waiting list or is already registered before showing the appropriate message.